+++ src/whatsapp/repository/repository.manager.ts
@@ -102,7 +102,12 @@ export class RepositoryBroker {
       this.logger.verbose('creating store path: ' + storePath);
 
       const tempDir = join(storePath, 'temp');
+      const chatwootDir = join(storePath, 'chatwoot');
 
+      if (!fs.existsSync(chatwootDir)) {
+        this.logger.verbose('creating chatwoot dir: ' + chatwootDir);
+        fs.mkdirSync(chatwootDir, { recursive: true });
+      }
       if (!fs.existsSync(tempDir)) {
         this.logger.verbose('creating temp dir: ' + tempDir);
         fs.mkdirSync(tempDir, { recursive: true });