+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.4.2 (2023-07-24 20:52)
+
+### Fixed
+
+* Fixed validation is set settings 
+* Adjusts in group validations
+* Ajusts in sticker message to chatwoot
+
 # 1.4.1 (2023-07-24 18:28)
 
 ### Fixed